requirements.txt
flask>=3.0
requests>=2.31
beautifulsoup4>=4.12
trafilatura>=1.6
sentence-transformers>=2.2
faiss-cpu>=1.7
groq>=0.4
python-dotenv>=1.0
